Fixes Eweka error
Some newsservers incorrectly return 'Article not found' when the article is available, but located on a different server. After a GROUP command, they redirect you to the right server, and if you try again, suddenly the article is found.

But Phuse did not always send a GROUP after 'Article not found', so this caused troubles with certain providers (like Eweka). This should be fixed now.